+++ legadilo/users/tests/test_commands/test_clean_users.py
@@ -1,13 +1,16 @@
 # SPDX-FileCopyrightText: 2023-2025 Legadilo contributors
 #
 # SPDX-License-Identifier: AGPL-3.0-or-later
+from datetime import timedelta
 
 import pytest
 import time_machine
 from allauth.account.models import EmailAddress
 from django.core.management import call_command
 
-from legadilo.users.models import User
+from legadilo.core.utils.time_utils import utcnow
+from legadilo.users import constants
+from legadilo.users.models import User, UserSession
 from legadilo.users.tests.factories import UserFactory
 
 
@@ -25,3 +28,42 @@ class TestCleanUsersCommand:
         call_command("clean_users")
 
         assert list(User.objects.all()) == [user_to_keep]
+
+    def test_clean_expired_sessions(self):
+        UserSession.objects.create(
+            session_key="expired_session",
+            user=UserFactory(),
+            expire_date=utcnow() - timedelta(days=100),
+            created_at=utcnow(),
+            updated_at=utcnow(),
+        )
+        UserSession.objects.create(
+            session_key="valid_session",
+            user=UserFactory(),
+            expire_date=utcnow() + timedelta(days=7),
+            created_at=utcnow(),
+            updated_at=utcnow(),
+        )
+
+        call_command("clean_users")
+
+        assert UserSession.objects.count() == 1
+        session = UserSession.objects.get()
+        assert session.session_key == "valid_session"
+
+    def test_clean_inactive_users(self, mocker):
+        send_mail_mock = mocker.patch("legadilo.users.models.user.send_mail")
+        UserFactory(
+            email="user_to_notify_14_days@example.com",
+            last_login=utcnow()
+            - constants.INACTIVE_USERS_RETENTION
+            + constants.INACTIVE_USERS_NOTIFICATION_THRESHOLDS[1],
+        )
+        UserFactory(
+            email="to_delete@example.com", last_login=utcnow() - constants.INACTIVE_USERS_RETENTION
+        )
+
+        call_command("clean_users")
+
+        assert send_mail_mock.call_count == 1
+        assert User.objects.count() == 1