+++ CHANGELOG.md
@@ -15,6 +15,11 @@ SPDX-License-Identifier: CC-BY-SA-4.0
 - Improve the display of reading lists selector.
 - Cleanup navigation in user settings.
 - Improve save actions in admin.
+- Rework how user sessions are managed to link them to the user.
+  - Sessions are now cleaned as part of the `clean_users` command. `clearsessions` cannot be used anymore.
+- Enable account deletion in the account parameters.
+- Inactive accounts are deleted after several notification emails. Both the deletion and the notification are done by the `clean_users` command.
+- Allow user to specify their language. Use the browser one by default and English as fallback.
 
 ## 25.12.5